Create HST_IDECode
This code should be run on an IDE to ensure everything works as it is the backbone for the Raspberry Pi version of this code. The only difference between this version and the Pi version is that the Pi version code has the addition of the RPi.GPIO library function calls being made in place of printing the 'container' arrays in the num1 and num2 functions.